src/aws/aios3/requirements.txt
python-dotenv

aioboto3